# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Item {
  id: string;
  name: string;
  type: string;
  price: number;
  unit: string;
  createdAt: string;
}

export interface ItemInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  type: string;
  /** @minimum 0 */
  price: number;
  /** @minLength 1 */
  unit: string;
}

export interface Entry {
  id: string;
  issuedAt: string;
  person: string;
  itemId: string;
  itemName: string;
  /** @minimum 1 */
  quantity: number;
  /** @minimum 0 */
  unitPrice: number;
  remarks: string;
  /** @nullable */
  tagSymbol: string | null;
  /** @nullable */
  tagColor: string | null;
}

export interface EntryInput {
  person: string;
  itemId: string;
  /** @minimum 1 */
  quantity: number;
  remarks: string;
  /** @nullable */
  tagSymbol: string | null;
  /** @nullable */
  tagColor: string | null;
}

export interface PersonTotal {
  person: string;
  quantity: number;
  total: number;
  lastIssuedAt: string;
}

export interface ReportSummary {
  totalQuantity: number;
  totalPrice: number;
  peopleCount: number;
  catalogCount: number;
  peopleTotals: PersonTotal[];
}

export interface Archive {
  id: string;
  label: string;
  archivedAt: string;
  totalQuantity: number;
  totalPrice: number;
  entries: Entry[];
}

/**
 * Resource not found
 */
export type NotFoundResponse = Error;