# lib/api-zod/src/generated/types/archive.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Entry } from './entry';

export interface Archive {
  id: string;
  label: string;
  archivedAt: Date;
  totalQuantity: number;
  totalPrice: number;
  entries: Entry[];
}