# lib/api-zod/src/generated/types/entry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Entry {
  id: string;
  issuedAt: Date;
  person: string;
  itemId: string;
  itemName: string;
  /** @minimum 1 */
  quantity: number;
  /** @minimum 0 */
  unitPrice: number;
  remarks: string;
  /** @nullable */
  tagSymbol: string | null;
  /** @nullable */
  tagColor: string | null;
}