# lib/api-zod/src/generated/types/entryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EntryInput {
  person: string;
  itemId: string;
  /** @minimum 1 */
  quantity: number;
  remarks: string;
  /** @nullable */
  tagSymbol: string | null;
  /** @nullable */
  tagColor: string | null;
}