# lib/api-zod/src/generated/types/itemInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ItemInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  type: string;
  /** @minimum 0 */
  price: number;
  /** @minLength 1 */
  unit: string;
}