# lib/api-zod/src/generated/types/reportSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PersonTotal } from './personTotal';

export interface ReportSummary {
  totalQuantity: number;
  totalPrice: number;
  peopleCount: number;
  catalogCount: number;
  peopleTotals: PersonTotal[];
}